♻ Simplify task
