+++ corehq/apps/data_interfaces/tests/test_tasks.py
@@ -40,40 +40,20 @@ class TestTasks(TestCase):
         self.assertEqual(mock_operate_on_payloads.call_count, 1)
         mock_operate_on_payloads.assert_called_with(['payload_id'], 'test_domain', 'test_action', mock_task)
 
-    @patch('corehq.apps.data_interfaces.tasks.generate_ids_and_operate_on_payloads')
-    def test_task_generate_ids_and_operate_on_payloads_no_action(self, mock_generate_ids):
-        with patch('corehq.apps.data_interfaces.tasks._') as _:
-            response = task_generate_ids_and_operate_on_payloads(
-                query_string_dict=['payload_id'],
-                domain='test_domain',
-                action=''
-            )
-            expected_response = {'messages': {'errors': [_('No action specified')]}}
-
-        self.assertEqual(response, expected_response)
-        self.assertEqual(mock_generate_ids.call_count, 0)
-
-    @patch('corehq.apps.data_interfaces.tasks.generate_ids_and_operate_on_payloads')
-    def test_task_generate_ids_and_operate_on_payloads_no_data(self, mock_generate_ids):
-        with patch('corehq.apps.data_interfaces.tasks._') as _:
-            response = task_generate_ids_and_operate_on_payloads(
-                query_string_dict=[],
-                domain='test_domain',
-                action=''
-            )
-            expected_response = {'messages': {'errors': [_('No data is supplied')]}}
-
-        self.assertEqual(response, expected_response)
-        self.assertEqual(mock_generate_ids.call_count, 0)
-
-    @patch('corehq.apps.data_interfaces.tasks.generate_ids_and_operate_on_payloads')
-    @patch('corehq.apps.data_interfaces.tasks.task_generate_ids_and_operate_on_payloads')
-    def test_task_generate_ids_and_operate_on_payloads_valid(self, mock_task, mock_generate_ids):
-        task_generate_ids_and_operate_on_payloads(
+    def test_task_generate_ids_and_operate_on_payloads_no_action(self):
+        response = task_generate_ids_and_operate_on_payloads(
             query_string_dict=['payload_id'],
             domain='test_domain',
-            action='test_action',
+            action=''
         )
+        self.assertEqual(response,
+                         {'messages': {'errors': ['No action specified']}})
 
-        self.assertEqual(mock_generate_ids.call_count, 1)
-        mock_generate_ids.assert_called_with(['payload_id'], 'test_domain', 'test_action', mock_task)
+    def test_task_generate_ids_and_operate_on_payloads_no_data(self):
+        response = task_generate_ids_and_operate_on_payloads(
+            query_string_dict=[],
+            domain='test_domain',
+            action=''
+        )
+        self.assertEqual(response,
+                         {'messages': {'errors': ['No data is supplied']}})
